Reorganized the SvelteApplicationMixin. Changed the sheet lock option to a static field.
Moved application interfaces to their own types file, to be expanded later 😇

Removed unused code.
